schelude-backend/telegram_bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
